backend/requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
motor==3.4.0
pymongo==4.7.2
python-dotenv==1.0.1
pydantic==2.7.1
passlib[bcrypt]==1.7.4
pyjwt==2.8.0
httpx==0.27.0
apscheduler==3.10.4
rembg==2.0.57
onnxruntime==1.18.0
Pillow==10.3.0
python-multipart==0.0.9
firebase-admin==6.4.0
google-cloud-storage==2.14.0
